bugs fixed

+++ AcoPlaceBackend/src/main/java/put/poznan/AcoPlaceBackend/controller/AnnouncementController.java
@@ -28,7 +28,7 @@ public class AnnouncementController {
     }
 
     @GetMapping("/announcement/{id}")
-    public Announcement getAdvertisementById(@PathVariable long id){
+    public Announcement getAdvertisementById(@PathVariable Integer id){
         return announcementService.getAnnouncementById(id);
     }
